requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-heroku==0.3.1
django-nose==1.4.7
djangorestframework==3.12.4
gunicorn==20.1.0
nose==1.3.7
pytest==6.2.4
pytest-django==4.4.0
pytest-subtests==0.5.0
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.5
